fix(android): ColumnScope receiver and ULong boundary conversion in director UI
Two compile errors from the UDL-sync commit (f26eaa6) that CI masked
behind the Gradle verification failure:

- DirectorNowScreen.kt: DirectorPanel's content lambda used `Column`
  (a composable function) as receiver type; the receiver is
  `ColumnScope`.
- QuorumMobileApp.kt: domain snapshot.version is Long; the elvis
  fallback `0uL` produced Comparable<*> against the bridge's ULong
  parameter. Keep the domain value Long and convert with .toULong()
  at the bridge call, matching the DirectorBridgeMapping boundary
  convention.

+++ apps/marquee/quorum-sense/android/app/src/main/java/se/reflective/quorum/ui/QuorumMobileApp.kt
@@ -80,8 +80,8 @@ private fun DirectorRootScreen(
         while (isActive) {
             snapshot = bridge.currentDirectorSnapshot()
             snapshotSource = bridge.directorSnapshotSource()
-            val version = snapshot?.version ?: 0uL
-            val updated = bridge.waitDirectorUpdate(version, 30_000u)
+            val version = snapshot?.version ?: 0L
+            val updated = bridge.waitDirectorUpdate(version.toULong(), 30_000u)
             if (!updated) {
                 // Poll again after timeout so fixture-only builds still refresh on intent.
             }
